Added error dialog if TTS goes wrong.
It is very unlikely we’ll ever get this from the current TTSClient 
objects, but the mechanism is there for later. On iOS/OSX, we don’t ever
report failure. On Android, we only report failure *to queue utterance*
which we probably won’t get even if voices aren’t installed. I have some
WIP on checking for voices, but it’s not behaving as I expected on an 
older tablet. TODO!