Reading obf data
Updates to read_json to read data from an open board format file.
Also a new index_obf.html which instructs read_json to read an obf file and index_ovf.html which is maintained for backward compatibility.

Note that at this point there are issues with dispalying obf files so this checkin isn't functional.